+++ micro-service-1/src/main/java/com/ebanking/controller/agentController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.view.RedirectView;
 
+import com.ebanking.config.Authentication;
 import com.ebanking.model.Agence;
 import com.ebanking.model.Agent;
 import com.ebanking.repository.AgenceRepository;
@@ -22,53 +23,69 @@ public class agentController {
 	@Autowired
 	private AgentRepository agentRepository;
 
+	@Autowired
+	private Authentication authenticated;
+
 	@GetMapping("/agent")
 	public String agent(Model model, @RequestParam(name = "idagence", defaultValue = "1") Long idAgence,
 			@RequestParam(name = "isModifier", defaultValue = "false") boolean isModifier) {
+		if (authenticated.isAuthenticated() == true) {
+			Agence agence = agenceRepository.findById(idAgence).get();
+			model.addAttribute("agence", agence);
+			model.addAttribute("nvAgent", new Agent());
+			model.addAttribute("isModifier", isModifier);
+			return "agent";
+		} else {
+			return "login";
+		}
 
-		Agence agence = agenceRepository.findById(idAgence).get();
-		model.addAttribute("agence", agence);
-		model.addAttribute("nvAgent", new Agent());
-		model.addAttribute("isModifier", isModifier);
-		return "agent";
 	}
 
 	@PostMapping("/creerAgent")
 	public RedirectView creerAgent(Model model, @ModelAttribute("nvAgent") Agent ag,
 			@RequestParam(name = "idagence", defaultValue = "1") Long idAgence,
 			@RequestParam(name = "idagent", defaultValue = "0") Long idAgent) {
-		Agence agence = agenceRepository.findById(idAgence).get();
-		if (idAgent != null && idAgent != 0) {
-			Agent agent = agentRepository.findById(idAgent).get();
-			agent.setCode(idAgent);
-			agent.setNom(ag.getNom());
-			agent.setPrenom(ag.getPrenom());
-			agent.setEmail(ag.getEmail());
-			agent.setCin(ag.getCin());
-			agent.setTelephone(ag.getTelephone());
-			agent.setAgence(agence);
-			agent.setVille(ag.getVille());
+		if (authenticated.isAuthenticated() == true) {
+			Agence agence = agenceRepository.findById(idAgence).get();
+			if (idAgent != null && idAgent != 0) {
+				Agent agent = agentRepository.findById(idAgent).get();
+				agent.setCode(idAgent);
+				agent.setNom(ag.getNom());
+				agent.setPrenom(ag.getPrenom());
+				agent.setEmail(ag.getEmail());
+				agent.setCin(ag.getCin());
+				agent.setTelephone(ag.getTelephone());
+				agent.setAgence(agence);
+				agent.setVille(ag.getVille());
 
-			agenceRepository.save(agence);
+				agenceRepository.save(agence);
+			} else {
+				ag.setAgence(agence);
+				agentRepository.save(ag);
+			}
+			RedirectView redirect = new RedirectView("agent");
+			redirect.addStaticAttribute("idagence", idAgence);
+			return redirect;
 		} else {
-			ag.setAgence(agence);
-			agentRepository.save(ag);
+			return new RedirectView("login");
 		}
-		RedirectView redirect = new RedirectView("agent");
-		redirect.addStaticAttribute("idagence", idAgence);
-		return redirect;
+
 	}
 
 	@GetMapping("/deletAgent")
 	public RedirectView deletAgent(Model model, @RequestParam(name = "idagent") Long idAgent) {
+		if (authenticated.isAuthenticated() == true) {
+			Agent agent = agentRepository.findById(idAgent).get();
+			agentRepository.delete(agent);
 
-		Agent agent = agentRepository.findById(idAgent).get();
-		agentRepository.delete(agent);
+			RedirectView redirect = new RedirectView("agent");
+			redirect.addStaticAttribute("idagence", agent.getAgence().getId());
+			redirect.addStaticAttribute("isModifier", true);
+			return redirect;
+		} else {
+			return new RedirectView("login");
+		}
 
-		RedirectView redirect = new RedirectView("agent");
-		redirect.addStaticAttribute("idagence", agent.getAgence().getId());
-		redirect.addStaticAttribute("isModifier", true);
-		return redirect;
 	}
 
 	@GetMapping("/signalerAgent")
